Fix bug in SKFunction: remove duplicate RequestSettings objects (#21)
### Motivation and Context
SKFunction offers 2 ways to update semantic functions' request settings.
The 2 APIs use different instances of `CompleteRequestSettings` instead
of one, leading to situations where the settings are out of sync or
settings requested by the user are ignored.

### Description
Remove the duplicate object and point the 2 APIs to the same settings
instance.
